async_api/service/requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.20.0
django-split-settings==1.1.0
django-debug-toolbar==3.4.0
pytest==7.1.2
pytest-django==4.5.2
isort==5.10.1
psycopg2-binary
django-cors-headers==3.13.0
wheel==0.37.1
uwsgi==2.0.20
elasticsearch==8.3.1
pydantic==1.9.1
psycopg2-binary
python-dotenv==0.20.0